Made test for boardGenerator
Made Test folder and created dummy test for the board generator.
Checks that if the json file has just "f "s then the board is filled with FloorTiles

Given a height and a width in a Json file, checks that the board generator returns a board with the same heigth and with.

Added JSon files used in the tests in folder "test/assets"